Improve input validation: Warn if a parameter results in values lower than a single sample
This may help a user to prevent cases where a combination of parameters and sampling rate
would result in events shorter than a sample that would not be automatically discarded as
implausible.
This was the case in #28, where a user reported errors with remodnav on data acquired with
a sampling rate of 60Hz. Because the minimal saccade duration of 0.01 seconds was shorter
than a single sample, 'saccade' events of zero length were passed to numpy functions instead
of skipped, which then resulted in unexpected crashed.

While the log message that would be issues with such a parametrization would not prevent errors,
it can help users to resolve them. This is less intrusive than replacing the values, which
may have a solid reason for existence and do not in all cases result in crashes